requirements.txt
Flask==3.1.2
gunicorn==21.2.0
numpy==1.26.4
pycryptodome==3.18.0
